(feat) Ensure nvim socket file consistency
If a nvim is closed in a bad way and the server socket file is not
deleted the next time a vmux session is started if the same name nvim
won't be able to stablish the connection. By deleting the socket file if
it's still there we avoid this issue.